# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API completa para gestão de equinos com Tokenização RWA (Real World Assets). Inclui: autenticação, gestão de equinos, tokenização, leilões, exames laboratoriais, nutrição com IA, treinamento, rankings, financeiro e muito mais.",
        "title": "Equinoid API - Sistema Perfeito",
        "termsOfService": "https://equinoid.com/terms",
        "contact": {
            "name": "Equinoid Support",
            "url": "https://equinoid.com/support",
            "email": "suporte@equinoid.com"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://equinoid.com/license"
        },
        "version": "2.0"
    },
    "host": "seahorse-app-28du8.ondigitalocean.app",
    "basePath": "/api/v1",
    "paths": {
...
        "/nutricao/equino/{equinoid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o plano nutricional de um equino",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrição"
                ],
                "summary": "Obter plano nutricional",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equinoid do equino",
                        "name": "equinoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/nutricao/equino/{equinoid}/ai-suggestion": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gera uma sugestão de plano nutricional usando inteligência artificial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrição"
                ],
                "summary": "Obter sugestão nutricional com IA",
                "parameters": [
                    {
                        "description": "Dados para sugestão",
                        "name": "sugestao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SugestaoIARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/nutricao/plano": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um novo plano nutricional para um equino",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrição"
                ],
                "summary": "Criar plano nutricional",
                "parameters": [
                    {
                        "description": "Dados do plano",
                        "name": "plano",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePlanoNutricionalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/nutricao/refeicoes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma refeição realizada pelo equino",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrição"
                ],
                "summary": "Criar registro de refeição",
                "parameters": [
                    {
                        "description": "Dados da refeição",
                        "name": "refeicao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRefeicaoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.AlimentoRefeicao": {
            "type": "object",
            "required": [
                "nome",
                "quantidade",
                "unidade"
            ],
            "properties": {
                "calorias": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "quantidade": {
                    "type": "number"
                },
                "unidade": {
                    "type": "string"
                }
            }
        },
...
        "models.CreatePlanoNutricionalRequest": {
            "type": "object",
            "required": [
                "equinoid",
                "objetivo",
                "peso_atual",
                "peso_ideal",
                "tipo_plano"
            ],
            "properties": {
                "equinoid": {
                    "type": "string"
                },
                "frequencia_refeicoes": {
                    "type": "integer",
                    "maximum": 6,
                    "minimum": 2
                },
                "gerar_com_ia": {
                    "type": "boolean"
                },
                "objetivo": {
                    "type": "string"
                },
                "observacoes_gerais": {
                    "type": "string"
                },
                "peso_atual": {
                    "type": "number"
                },
                "peso_ideal": {
                    "type": "number"
                },
                "restricoes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "suplementos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tipo_plano": {
                    "$ref": "#/definitions/models.TipoPlano"
                }
            }
        },
...
        "models.CreateRefeicaoRequest": {
            "type": "object",
            "required": [
                "alimentos",
                "equinoid",
                "tipo_refeicao"
            ],
            "properties": {
                "alimentos": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.AlimentoRefeicao"
                    }
                },
                "data_refeicao": {
                    "type": "string"
                },
                "equinoid": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "tipo_refeicao": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.SugestaoIARequest": {
            "type": "object",
            "required": [
                "equinoid",
                "objetivo",
                "peso_atual",
                "tipo_plano"
            ],
            "properties": {
                "equinoid": {
                    "type": "string"
                },
                "objetivo": {
                    "type": "string"
                },
                "peso_atual": {
                    "type": "number"
                },
                "tipo_plano": {
                    "$ref": "#/definitions/models.TipoPlano"
                }
            }
        },
...
        "models.TipoPlano": {
            "type": "string",
            "enum": [
                "manutencao",
                "ganho_massa",
                "perda_peso",
                "recuperacao",
                "alto_rendimento",
                "gestacao",
                "lactacao"
            ],
            "x-enum-varnames": [
                "TipoPlanoManutencao",
                "TipoPlanoGanhoMassa",
                "TipoPlanoPerdaPeso",
                "TipoPlanoRecuperacao",
                "TipoPlanoAltoRendimento",
                "TipoPlanoGestacao",
                "TipoPlanoLactacao"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}